split dynamic test recipes for clarity, update README
